Update for Windows

+++ devtool/vim/vim_config.ps1
@@ -10,8 +10,10 @@
 Write-Host -NoNewline " => Config "
 Write-Host -ForegroundColor YELLOW "neovim"
 If (Get-Command nvim -errorAction SilentlyContinue) {
-	Remove-Item "C:\Users\$($env:UserName)\.config\nvim\pack\bundle\start\fzf"
-	git clone https://github.com/junegunn/fzf "C:\Users\$($env:UserName)\.config\nvim\pack\bundle\start"
+	If (Test-Path "C:\Users\$($env:UserName)\.config\nvim\pack\bundle\start\fzf") {
+		Remove-Item "C:\Users\$($env:UserName)\.config\nvim\pack\bundle\start\fzf"
+	}
+	git clone https://github.com/junegunn/fzf "C:\Users\$($env:UserName)\.config\nvim\pack\bundle\start\fzf"
 } Else {
 	Write-Host "    Not installed."
 }
